rename build_skia to skia
