+++ skia-bindings/build_support.rs
@@ -3,7 +3,7 @@
 pub mod android;
 pub mod binaries_config;
 pub mod bind_skia;
-pub mod build_skia;
+pub mod skia;
 pub mod cargo;
 pub mod clang;
 pub mod features;
